requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.1
